% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTDAcif.R
\name{DTDAcif}
\alias{DTDAcif}
\title{Doubly Truncated Data Analysis, Cumulative Incidence Function}
\usage{
DTDAcif(x, u, v, comp.event = z, method = c("indep", "dep"))
}
\arguments{
\item{x}{Numeric vector corresponding the variable of ultimate interest.}

\item{u}{Numeric vector corresponding to the left truncation variable.}

\item{v}{TODO}

\item{comp.event}{TODO}

\item{method}{TODO}

\item{boot}{TODO}

\item{B}{TODO}

\item{level}{TODO}
}
\value{
A list containing:
}
\description{
TODO
}
\section{Acknowledgements}{

\itemize{
\item{Jacobo de Uña-Álvarez was supported by Grant MTM2014-55966-P, Spanish Ministry of Economy and Competitiveness.}
\item{José Carlos Soage was supported by Red Tecnológica de Matemática Industrial (Red TMATI), Cons. de Cultura, Educación e OU, Xunta de Galicia (ED341D R2016/051) and by Grupos de Referencia Competitiva, Consolidación y Estructuración de Unidades de Investigación Competitivas del SUG, Cons. de Cultura, Educación e OU, Xunta de Galicia (GRC ED431C 2016/040).}
}
}

\examples{


}
\references{
TODO
}
\author{
\itemize{
\item{de Uña-Álvarez, Jacobo.}
\item{Soage González, José Carlos.}
\item{Maintainer: José Carlos Soage González. \email{jsoage@uvigo.es}}
}
}
