% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DTDAcif.R
\name{plot.DTDAcif}
\alias{plot.DTDAcif}
\title{plot.DTDAcif}
\usage{
\method{plot}{DTDAcif}(x, intervals = FALSE, level = 0.95, main = "",
  xlab = "", ylab = "", ylim, xlim, ...)
}
\arguments{
\item{x}{DTDAcif object.}

\item{intervals}{Logical. If TRUE confidence intervals are calculated if standard deviation was calculated before.}

\item{level}{Confidence level of the standard deviation of the cifs. Default is 0.95.}

\item{main}{An overall title for the plot.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{ylim}{Limit over the y axis.}

\item{xlim}{Limit over the x axis.}

\item{...}{Additional parameters.}
}
\description{
S3 method to plot a DTDAcif object by using the generic plot function.
}
\author{
\itemize{
\item{de Uña-Álvarez, Jacobo.}
\item{Soage González, José Carlos.}
\item{Maintainer: José Carlos Soage González. \email{jsoage@uvigo.es}}
}
}
