% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DTDAcif.R
\name{summary.DTDAcif}
\alias{summary.DTDAcif}
\title{summary.DTDAcif}
\usage{
\method{summary}{DTDAcif}(object, ...)
}
\arguments{
\item{object}{DTDAcif object.}

\item{...}{Additonal parameters.}
}
\description{
S3 method to summarize a DTDAcif object by using the generic summary function.
}
\author{
\itemize{
\item{de Uña-Álvarez, Jacobo.}
\item{Soage González, José Carlos.}
\item{Maintainer: José Carlos Soage González. \email{jsoage@uvigo.es}}
}
}
